+++ app/src/androidTest/java/nextstep/signup/fixtures/TestFixture.kt
@@ -3,7 +3,7 @@ package nextstep.signup.fixtures
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import nextstep.signup.ValidationResult
+import nextstep.signup.ui.model.ValidationResult
 
 @Composable
 fun FakeSignUpField(
