+++ app/src/main/java/nextstep/signup/ui/model/ValidationResult.kt
@@ -1,3 +1,3 @@
-package nextstep.signup
+package nextstep.signup.ui.model
 
 data class ValidationResult(val isValid: Boolean, val errorMessage: String = "")
